requirements.txt
telebot
pyTelegramBotAPI
gTTS
googletrans==3.1.0a0
pydub
langdetect
fasttext
vosk
wave
ffmpeg-python